+++ API/Features/Reservations/Customers/Dtos/CustomerWriteDto.cs
@@ -10,7 +10,7 @@ namespace API.Features.Reservations.Customers {
         public int NationalityId { get; set; }
         public int TaxOfficeId { get; set; }
         // Fields
-        public int VatPercent { get; set; }
+        public decimal VatPercent { get; set; }
         public int VatPercentId { get; set; }
         public int VatExemptionId { get; set; }
         public string Description { get; set; }
