+++ API/Features/Reservations/ShipOwners/Models/ShipOwner.cs
@@ -14,7 +14,7 @@ namespace API.Features.Reservations.ShipOwners {
         public int NationalityId { get; set; }
         public int TaxOfficeId { get; set; }
         // Fields
-        public int VatPercent { get; set; }
+        public decimal VatPercent { get; set; }
         public int VatPercentId { get; set; }
         public int VatExemptionId { get; set; }
         public string Description { get; set; }
